Change Proposed Date to accept US Date Format
Depending what format a date was entered in, Proposed Date was losing
values because the date string would not parse or would parse the wrong
values because it was parsing the string incorrectly. ActiveRecord was
storing '12/4/2019' as '2019-04-12' and wouldn't store '12/16/2019' at
all because 2019-16-12 is not a valid date.

Because we wanted to maintain the ability to enter US dates and also use
a date picker, we have modified the bendyworks copy of best_in_place so
that it supports the use of HTML5 `<input type='date'>`, removed the
jquery datepicker integration on the field, and now rely on the browser
to localize the value of the date field, and use BIP's display_as to
display the correct format once the date saves on the backend.

This caused problems for Selenium, as it will not allow entry of a date
that does not conform to ISO8601, and caused problems for our sorting
tests because with the US date format the dates no longer sort in the
same manner as strings would sort. The tests were updated to deal with
these two issues.

Co-authored-by: Jacob Ela <jacob@bendyworks.com>
Co-authored-by: Amy Terry-Penak <amy@bendyworks.com>